add feature vis
